fix(tui): share chrome with execute
