+++ src/tui.rs
@@ -1,1 +1,7 @@
 pub(crate) mod chrome;
+pub(crate) mod terminal;
+
+pub(crate) use chrome::Chrome;
+#[cfg(test)]
+pub(crate) use terminal::compact_viewport_height;
+pub(crate) use terminal::run_scrollable_text;
